Updated schema to match Postgres

+++ src/schema.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct User {
     name: String,
-    age: u8,
+    age: i32,
 }
